Refactor: move models and enums to PPMModelLibrary
