+++ DataAccess/DataAccess.csproj
@@ -31,7 +31,6 @@
 
   <ItemGroup>
     <Folder Include="Migrations\" />
-    <Folder Include="Enums\" />
   </ItemGroup>
 
 </Project>
