+++ PPMModelLibrary/Models/Users/Administrator/Administrator.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace PPMModelLibrary.Models.Users.Administrator
+{
+    public class Administrator : IAdministrator
+    {
+        [Key]
+        public int Id { get; set; }
+        public ApplicationUser User { get; set; }
+    }
+}
