+++ PPMModelLibrary/Models/Users/Administrator/IAdministrator.cs
@@ -1,0 +1,6 @@
+﻿namespace PPMModelLibrary.Models.Users.Administrator
+{
+    public interface IAdministrator : IUser
+    {
+    }
+}
