+++ PPMModelLibrary/Models/Users/IUser.cs
@@ -1,0 +1,8 @@
+﻿namespace PPMModelLibrary.Models.Users
+{
+    public interface IUser
+    {
+        public int Id { get; set; }
+        public ApplicationUser User { get; set; }
+    }
+}
