+++ PPMModelLibrary/Models/Users/Owner/IOwner.cs
@@ -1,0 +1,6 @@
+﻿namespace PPMModelLibrary.Models.Users.Owner
+{
+    public interface IOwner : IUser
+    {
+    }
+}
