+++ PPMModelLibrary/Models/Users/Owner/Owner.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace PPMModelLibrary.Models.Users.Owner
+{
+    public class Owner : IOwner
+    {
+        [Key]
+        public int Id { get; set; }
+        public ApplicationUser User { get; set; }
+    }
+}
