+++ PPMModelLibrary/Models/Users/Tenant/ITenant.cs
@@ -1,0 +1,6 @@
+﻿namespace PPMModelLibrary.Models.Users.Tenant
+{
+    public interface ITenant : IUser
+    {
+    }
+}
