+++ PPMModelLibrary/Models/Users/Tenant/Tenant.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace PPMModelLibrary.Models.Users.Tenant
+{
+    public class Tenant : ITenant
+    {
+        [Key]
+        public int Id { get; set; }
+        public ApplicationUser User { get; set; }
+    }
+}
